+++ src/backend/cpu/CpuWorker.cpp
@@ -361,7 +361,9 @@ void xmrig::CpuWorker<N>::start()
             }
         }
 
-        consumeJob();
+        if (!Nonce::isPaused()) {
+            consumeJob();
+        }
     }
 }