+++ src/base/kernel/interfaces/IConfig.h
@@ -176,6 +176,7 @@ public:
         CCWorkerId              = 9008,
         CCUpdateInterval        = 9009,
         CCRetriesToFailover     = 9010,
+        CCProxyServer           = 9011,
 
         // xmrigCC added params
         ForceAutoconfigKey   = 9100,