+++ src/core/config/Config_platform.h
@@ -158,6 +158,7 @@ static const option options[] = {
     { "cc-reboot-cmd",         1, nullptr, IConfig::CCRebootCmd           },
     { "cc-url",                1, nullptr, IConfig::CCUrl                 },
     { "cc-access-token",       1, nullptr, IConfig::CCAccessToken         },
+    { "cc-http-proxy",         1, nullptr, IConfig::CCProxyServer         },
     { "cc-worker-id",          1, nullptr, IConfig::CCWorkerId            },
     { "cc-update-interval-s",  1, nullptr, IConfig::CCUpdateInterval      },
     { "cc-retries-to-failover",1, nullptr, IConfig::CCRetriesToFailover   },